third/core/requirements.txt
Django==3.1
djangorestframework==3.11.1
drf-yasg==1.20.0
